+++ D20Tek.Spectre.Console.Extensions.UnitTests/Mocks/MockCommand.cs
@@ -1,22 +1,15 @@
 ﻿//---------------------------------------------------------------------------------------------------------------------
 // Copyright (c) d20Tek.  All rights reserved.
 //---------------------------------------------------------------------------------------------------------------------
-using Microsoft.Extensions.DependencyInjection;
-using Moq;
 using Spectre.Console.Cli;
 
 namespace D20Tek.Spectre.Console.Extensions.UnitTests.Mocks
 {
-    internal class MockDIStartup : StartupBase<IServiceCollection>
+    internal class MockCommand : Command
     {
-        public override ITypeRegistrar ConfigureServices(IServiceCollection services)
+        public override int Execute(CommandContext context)
         {
-            return new Mock<ITypeRegistrar>().Object;
-        }
-
-        public override IConfigurator ConfigureCommands(IConfigurator config)
-        {
-            return config;
+            return 0;
         }
     }
 }