+++ D20Tek.Spectre.Console.Extensions.UnitTests/Mocks/MockStartup.cs
@@ -1,17 +1,15 @@
 ﻿//---------------------------------------------------------------------------------------------------------------------
 // Copyright (c) d20Tek.  All rights reserved.
 //---------------------------------------------------------------------------------------------------------------------
-using Moq;
-using Ninject;
 using Spectre.Console.Cli;
 
 namespace D20Tek.Spectre.Console.Extensions.UnitTests.Mocks
 {
-    internal class MockNinjectStartup : StartupBase<StandardKernel>
+    internal class MockStartup : StartupBase
     {
-        public override ITypeRegistrar ConfigureServices(StandardKernel kernel)
+        public override void ConfigureServices(ITypeRegistrar registrar)
         {
-            return new Mock<ITypeRegistrar>().Object;
+            registrar.Register(typeof(IMockService), typeof(MockService));
         }
 
         public override IConfigurator ConfigureCommands(IConfigurator config)