+++ samples/SimpleInjector.Cli/Program.cs
@@ -4,14 +4,14 @@
 using D20Tek.Samples.Common.Commands;
 using D20Tek.Spectre.Console.Extensions;
 
-namespace DependencyInjection.Cli
+namespace SimpleInjector.Cli
 {
     public class Program
     {
         public static async Task<int> Main(string[] args)
         {
             return await new CommandAppBuilder()
-                             .WithDIContainer()
+                             .WithSimpleInjectorContainer()
                              .WithStartup<Startup>()
                              .WithDefaultCommand<DefaultCommand>()
                              .Build()